# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: prayer-times
    description: Swedish prayer timetable operations
  - name: reminders
    description: Device prayer push subscriptions
paths:
  /reminders/device:
    put:
      operationId: saveReminderDevice
      tags: [reminders]
      summary: Replace a device's reminder subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReminderDeviceInput"
      responses:
        "200":
          description: Subscription saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReminderDeviceStatus"
        "400":
          description: Invalid subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /prayer-times:
    get:
      operationId: getPrayerTimes
      tags: [prayer-times]
      summary: Get Islamiska förbundet prayer times
      description: Proxies the current Swedish city timetable dataset published by salat.muslim.se and sourced from Islamiska förbundet.
      responses:
        "200":
          description: Current Swedish prayer timetable dataset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerTimesDataset"
        "502":
          description: The upstream timetable is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    ReminderDeviceInput:
      type: object
      required: [token, enabled, city, mode, prayers, localUntil]
      properties:
        token: { type: string, pattern: "^Expo(nent)?PushToken\\[[A-Za-z0-9_-]+\\]$", maxLength: 256 }
        enabled: { type: boolean }
        city: { type: string, maxLength: 120 }
        mode: { type: string, enum: [vibration, sound] }
        prayers:
          type: object
          required: [Fajr, Dhuhr, Asr, Maghrib, Isha]
          properties:
            Fajr: { type: boolean }
            Dhuhr: { type: boolean }
            Asr: { type: boolean }
            Maghrib: { type: boolean }
            Isha: { type: boolean }
        localUntil: { type: string, format: date-time }
    ReminderDeviceStatus:
      type: object
      required: [active]
      properties:
        active: { type: boolean }
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PrayerTimesDataset:
      type: object
      additionalProperties: true
      required: [source, downloadedAt, cities, tables]
      properties:
        source:
          type: string
        downloadedAt:
          type: string
        cities:
          type: array
          items:
            type: object
            additionalProperties: true
        tables:
          type: object
          additionalProperties: true
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string