# lib/api-zod/src/generated/types/reminderDeviceInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReminderDeviceInputMode } from './reminderDeviceInputMode';
import type { ReminderDeviceInputPrayers } from './reminderDeviceInputPrayers';

export interface ReminderDeviceInput {
  /**
     * @maxLength 256
     * @pattern ^Expo(nent)?PushToken\[[A-Za-z0-9_-]+\]$
     */
  token: string;
  enabled: boolean;
  /** @maxLength 120 */
  city: string;
  mode: ReminderDeviceInputMode;
  prayers: ReminderDeviceInputPrayers;
  localUntil: Date;
}